+++ src/lib/applications/Custom.svelte
@@ -7,7 +7,7 @@
   export let Width = 640
   export let Height = 360
 
-  import Window from './Window.svelte'
+  import Window from '../Window.svelte'
 
 </script>
